feat: Validate by Header Response
Hello @colindickson I wanted to be able to validate the response header.
For example :
In a request I can get two different content-type dependning of some parameters.
And I want to a be able o validate the Headers I receive.

So I have coded a simple code that evaluate if the header is present, and if the value is the good one.